{{cookiecutter.project_slug}}/server/requirements.txt
aioredis
databases[postgresql]
python-dotenv
passlib
sqlalchemy
argon2-cffi
jwt
python-multipart